Solve it with React
