+++ hugo/content/playground/common.ts
@@ -4,19 +4,6 @@
  * terms of the MIT License, which is available in the project root.
  ******************************************************************************/
 
-import { compressToEncodedURIComponent } from "lz-string";
-import {
-  AbstractMessageReader,
-  DataCallback,
-  Diagnostic,
-  Disposable,
-  Emitter,
-  MessageReader,
-} from "vscode-languageserver";
-import {
-  BrowserMessageReader,
-  BrowserMessageWriter,
-} from "vscode-languageserver/browser";
 import {
   LangiumInitialContent,
   LangiumMonarchContent,
@@ -25,185 +12,10 @@ import {
 import { generateMonarch } from "./monarch-generator";
 import { createServicesForGrammar } from "langium/lib/grammar/grammar-util";
 import { decompressFromEncodedURIComponent } from "lz-string";
-import { AstNode, Reference } from "langium";
-import Mustache from "mustache";
-
-export { BrowserMessageReader, BrowserMessageWriter };
-
-declare type DedicatedWorkerGlobalScope = any;
-
-export type PlaygroundMessageType = "validated" | "changing" | "error" | "ast";
-
-export interface PlaygroundMessageBase {
-  type: PlaygroundMessageType;
-}
-
-export interface PlaygroundError extends PlaygroundMessageBase {
-  type: "error";
-  errors: Diagnostic[];
-}
-
-export interface PlaygroundOK extends PlaygroundMessageBase {
-  type: "validated";
-  grammar: string;
-}
-
-export interface PlaygroundAst extends PlaygroundMessageBase {
-  type: "ast";
-  root: AstNode;
-}
-
-export interface PlaygroundChanging extends PlaygroundMessageBase {
-  type: "changing";
-}
-
-export type PlaygroundMessage =
-  | PlaygroundChanging
-  | PlaygroundError
-  | PlaygroundOK
-  | PlaygroundAst;
-
-export interface MessageBase {
-  jsonrpc: "2.0";
-}
-
-export interface Request extends MessageBase {
-  method: string;
-  params?: any[];
-  id: string;
-}
-
-export interface ResponseOK extends MessageBase {
-  result: any;
-  id: string;
-}
-
-export interface ResponseError extends MessageBase {
-  error: any;
-  id: string;
-}
-
-export interface Notification extends MessageBase {
-  method: string;
-  params?: any[];
-}
-
-export interface PlaygroundParameters {
-  grammar: string;
-  content: string;
-}
-
-export function isNotification(msg: Message): msg is Notification {
-  return !msg["id"] && msg["method"];
-}
-
-type Message = Request | ResponseError | ResponseOK | Notification;
-
-const MagicAction = "PlaygroundMagic";
-
-export interface MessageWrapper<T> {
-  wrap(message: T): Message;
-  unwrap(message: Message): T | null;
-}
-
-export type MessageCallback<T> = (data: T) => void;
-
-export class PlaygroundWrapper implements MessageWrapper<PlaygroundMessage> {
-  wrap(message: PlaygroundMessage): Message {
-    return {
-      jsonrpc: "2.0",
-      method: MagicAction,
-      params: [message],
-    } as Notification;
-  }
-  unwrap(message: Message): PlaygroundMessage | null {
-    if (isNotification(message) && message.method === MagicAction) {
-      return message.params![0] as PlaygroundMessage;
-    }
-    return null;
-  }
-}
-
-export class ByPassingMessageReader<T>
-  extends AbstractMessageReader
-  implements MessageReader
-{
-  private _onData: Emitter<Message>;
-  private _onByPass: Emitter<T>;
-  private _messageListener: (event: MessageEvent) => void;
-
-  public constructor(
-    port: MessagePort | Worker | DedicatedWorkerGlobalScope,
-    wrapper: MessageWrapper<T>
-  ) {
-    super();
-    this._onData = new Emitter<Message>();
-    this._onByPass = new Emitter<T>();
-    this._messageListener = (event: MessageEvent) => {
-      const unwrapped = wrapper.unwrap(event.data);
-      if (unwrapped) {
-        this._onByPass.fire(unwrapped);
-        return;
-      }
-      this._onData.fire(event.data);
-    };
-    port.addEventListener("error", (event) => this.fireError(event));
-    port.onmessage = this._messageListener;
-  }
-
-  public listen(callback: DataCallback): Disposable {
-    return this._onData.event((x) => callback(x as any));
-  }
-
-  public listenByPass(callback: MessageCallback<T>): Disposable {
-    return this._onByPass.event(callback);
-  }
-}
-
-export class ByPassingMessageWriter<T> extends BrowserMessageWriter {
-  public constructor(
-    port: MessagePort | Worker | DedicatedWorkerGlobalScope,
-    private wrapper: MessageWrapper<T>
-  ) {
-    super(port);
-  }
-
-  public byPassWrite(message: T) {
-    return this.write(this.wrapper.wrap(message));
-  }
-}
-
-export interface MonacoConnection {
-  reader: ByPassingMessageReader<PlaygroundMessage>;
-  writer: ByPassingMessageWriter<PlaygroundMessage>;
-}
-
-export interface MonacoEditorResult {
-  out: ByPassingMessageReader<PlaygroundMessage>;
-  in: ByPassingMessageWriter<PlaygroundMessage>;
-  editor: MonacoClient;
-  overlay(visible: boolean): void;
-}
-
-export interface MonacoConfig {
-  getMainCode(): string;
-  setMainCode(code: string): void;
-  setMainLanguageId(name: string): void;
-  setMonarchTokensProvider(monarch: any): void;
-  theme: string;
-  useLanguageClient: boolean;
-  useWebSocket: boolean;
-  setMonacoEditorOptions(monacoEditorOptions: Record<string, unknown>): void;
-}
-
-export interface MonacoClient {
-  getEditorConfig(): MonacoConfig;
-  setWorker(worker: Worker, connection: MonacoConnection): void;
-  startEditor(domElement: HTMLElement): void;
-  updateLayout(): void;
-  dispose(): Promise<void>;
-  getMainCode(): string;
-}
+import { render } from "./Tree";
+import { PlaygroundMessage, PlaygroundMessageType, PlaygroundParameters } from "./types";
+import { ByPassingMessageReader, ByPassingMessageWriter, MonacoClient, MonacoEditorResult, PlaygroundWrapper } from "./monaco-utils";
+export { share } from "./utils";
 
 export function setupEditor(
   domElement: HTMLElement,
@@ -398,205 +210,11 @@ export function setupPlayground(
       content: dslContent,
     });
 
-    interface TypeNodeBase {
-      kind: "object" | "array" | "string" | "boolean" | "number" | 'reference';
-    }
-    interface SimpleType extends TypeNodeBase {
-      kind: "object";
-      $type: string;
-      properties: PropertyNode[];
-    }
-    interface ReferenceType extends TypeNodeBase {
-      kind: "reference";
-      $text: string;
-      type: TypeNode;
-    }
-    type PrimitiveMapping = {
-      string: string;
-      boolean: boolean;
-      number: number;
-    };
-    interface PrimitiveType<T extends "number" | "string" | "boolean">
-      extends TypeNodeBase {
-      kind: T;
-      value: PrimitiveMapping[T];
-    }
-
-    interface ArrayType extends TypeNodeBase {
-      kind: "array";
-      children: TypeNode[];
-    }
-
-    type TypeNode =
-      | SimpleType
-      | ArrayType
-      | PrimitiveType<"boolean">
-      | PrimitiveType<"number">
-      | PrimitiveType<"string">
-      | ReferenceType;
-
-    interface PropertyNode {
-      name: string;
-      type: TypeNode;
-    }
-
-    function preprocessAstNode(node: AstNode): SimpleType {
-      const properties: PropertyNode[] = Object.keys(node)
-        .filter((n) => !n.startsWith("$"))
-        .map((n) => {
-          const valueOrValues = node[n] as AstNode | AstNode[] | 'string' | 'number' | 'boolean' | Reference;
-          if (Array.isArray(valueOrValues)) {
-            return {
-              name: n,
-              type: preprocessArrayType(valueOrValues),
-            } as PropertyNode;
-          } else if(typeof valueOrValues === 'object') {
-            if('$refText' in valueOrValues) {
-              return {
-                name: n,
-                type: preprocessReferenceNode(valueOrValues),
-              } as PropertyNode; 
-            }
-            return {
-              name: n,
-              type: preprocessAstNode(valueOrValues),
-            } as PropertyNode;
-          } else if(typeof valueOrValues === 'string') {
-            return {
-              name: n,
-              type: {
-                kind: "string",
-                value: valueOrValues
-              }
-            } as PropertyNode;
-          } else if(typeof valueOrValues === 'number') {
-            return {
-              name: n,
-              type: {
-                kind: "number",
-                value: valueOrValues
-              }
-            } as PropertyNode;
-          } else {
-            return {
-              name: n,
-              type: {
-                kind: "boolean",
-                value: valueOrValues
-              }
-            } as PropertyNode;
-          }
-        });
-      return {
-        kind: "object",
-        $type: node.$type,
-        properties,
-      };
-    }
-
-    function preprocessReferenceNode(node: Reference<AstNode>): ReferenceType {
-      return {
-        kind: "reference",
-        $text: node.$refText,
-        type: preprocessAstNode(node.ref!)
-      }
-    }
-
-    function preprocessArrayType(nodes: AstNode[]): ArrayType {
-      const children = nodes.map(preprocessAstNode);
-      return {
-        kind: "array",
-        children,
-      };
-    }
-
-    function applyTemplate<T>(
-      templateId: string,
-      data: T
-    ) {
-      var template = document.getElementById(templateId)!.innerHTML;
-      return Mustache.render(template, data, undefined, {
-        tags: ["(%", "%)"],
-      });
-    }
-
-    function buildType(domElement: HTMLElement, root: TypeNode) {
-      debugger
-      switch(root.kind) {
-        case 'object': return buildSimpleType(domElement, root);
-        case 'boolean': return buildBoolean(domElement, root);
-        case 'string': return buildString(domElement, root);
-        case 'number': return buildNumber(domElement, root);
-       // case 'reference': return buildReference(domElement, root);
-        case 'array': return buildArray(domElement, root);
-      }
-    }
-
-    function buildArray(domElement: HTMLElement, root: ArrayType) {
-      domElement.innerHTML = applyTemplate("template-array", root);
-      const contentDom = (domElement.querySelector('.value-body') as HTMLElement)!;
-      for (const child of root.children) {
-        const li = document.createElement('li');
-        li.className = 'entry toggable closed';
-        buildType(li, child);
-        contentDom.appendChild(li);
-      }
-    }
-
-    function buildBoolean(domElement: HTMLElement, root: PrimitiveType<'boolean'>) {
-      domElement.innerHTML = applyTemplate("template-boolean", root);
-    }
-
-    function buildString(domElement: HTMLElement, root: PrimitiveType<'string'>) {
-      domElement.innerHTML = applyTemplate("template-string", root);
-    }
-
-    function buildNumber(domElement: HTMLElement, root: PrimitiveType<'number'>) {
-      domElement.innerHTML = applyTemplate("template-number", root);
-    }
-
-    function buildSimpleType(domElement: HTMLElement, root: SimpleType) {
-      let initializedChildren = false;
-      domElement.innerHTML = applyTemplate("template-simple", root);
-      const titleDom = domElement.querySelector(
-        ".type"
-      )! as HTMLElement;
-      const listItemDom = domElement.querySelector(
-        "li.entry"
-      )! as HTMLElement;
-      const valueBody = domElement.querySelector('.value-body')!;
-      titleDom.onclick = () => {
-        if(!listItemDom.classList.toggle("closed") && !initializedChildren) {
-          buildProperties(valueBody, root.properties);
-          initializedChildren = true;
-        }
-      };
-    }
-
-    function buildProperties(element: Element, properties: PropertyNode[]) {
-      for (const property of properties) {
-        const li = document.createElement('li');
-        li.className = 'entry toggable closed';
-        li.innerHTML = applyTemplate('template-property', property);
-        element.appendChild(li);
-
-        const body = li.querySelector('.property-body')!;
-        let initializedType = false;
-        (li as HTMLElement).onclick = () => {
-          if(!li.classList.toggle('closed') && !initializedType) {
-            buildType(body as HTMLElement, property.type);
-            initializedType = true;
-          }
-        };
-      }
-    }
-
     userDefined!.out.listenByPass((data) => {
       if (data.type !== "ast") {
         return;
       }
-      const preprocessed = preprocessAstNode(data.root);  
-      buildSimpleType(syntaxTreeDomElement, preprocessed);
+      render(data.root);
     });
   });
 
@@ -612,35 +230,3 @@ export function setupPlayground(
     } as PlaygroundParameters;
   };
 }
-
-export async function share(grammar: string, content: string): Promise<void> {
-  const compressedGrammar = compressToEncodedURIComponent(grammar);
-  const compressedContent = compressToEncodedURIComponent(content);
-  const url = new URL("/playground", window.origin);
-  url.searchParams.append("grammar", compressedGrammar);
-  url.searchParams.append("content", compressedContent);
-  await navigator.clipboard.writeText(url.toString());
-}
-
-export function throttle<T>(milliseconds: number, action: (input: T) => void) {
-  let timeout: NodeJS.Timeout | undefined = undefined;
-
-  function clear() {
-    if (timeout) {
-      clearTimeout(timeout);
-      timeout = undefined;
-    }
-  }
-
-  return {
-    clear,
-    call: (input: T) => {
-      clear();
-      timeout = setTimeout(() => {
-        action(input);
-      }, milliseconds);
-    },
-  };
-}
-
-
