+++ hugo/package.json
@@ -27,10 +27,12 @@
     "lz-string": "^1.4.4",
     "monaco-editor-workers": "0.34.2",
     "monaco-editor-wrapper": "1.1.0",
-    "vscode-languageserver": "8.0.2",
-    "@types/mustache": "^4.2.1"
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "vscode-languageserver": "8.0.2"
   },
   "dependencies": {
-    "mustache": "^4.2.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   }
 }
